requirements.txt
opencv-python>=4.9.0
numpy>=1.26.0
scipy>=1.12.0
pydantic>=2.6.0
pyyaml>=6.0.1
typer[all]>=0.12.0
